requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pytest
pytest-asyncio
python-dotenv
httpx